[api] Make comment class polymorphic
I introduced the migration and the relations needed to make the Comment
model polymorphic. I eliminated the CommentPackage, CommentProject and
CommentRequest models as they are not longer needed and I moved their
methods to the Comment model. I also removed the check_is_child
validation in Comment as it doesn't make sense now.